+++ cmd/del.go
@@ -3,37 +3,67 @@ package cmd
 import (
 	"bufio"
 	"fmt"
+	"io"
 	"os"
 
+	"github.com/chroju/parade/ssmctl"
 	"github.com/fatih/color"
 	"github.com/spf13/cobra"
 )
 
-var (
-	isForceDelete bool
-	// DelCommand is the command to delete key value
-	DelCommand = &cobra.Command{
-		Use:     "del <key>",
-		Short:   "Delete key and value in your parameter store.",
-		Args:    cobra.ExactArgs(1),
-		PreRunE: initializeCredential,
+type delOption struct {
+	Key           string
+	IsForceDelete bool
+
+	SSMManager ssmctl.SSMManager
+
+	Out    io.Writer
+	ErrOut io.Writer
+}
+
+func newDelCommand(globalOption *GlobalOption) *cobra.Command {
+	o := &delOption{}
+
+	cmd := &cobra.Command{
+		Use:   "del <key>",
+		Short: "Delete key and value in your parameter store.",
+		Args:  cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			return del(args)
+			o.SSMManager = globalOption.SSMManager
+			if o.SSMManager == nil {
+				ssmManager, err := initializeCredential(globalOption.Profile, globalOption.Region)
+				if err != nil {
+					return err
+				}
+				o.SSMManager = ssmManager
+			}
+
+			args = cmd.Flags().Args()
+			o.Key = args[0]
+
+			o.Out = globalOption.Out
+			o.ErrOut = globalOption.ErrOut
+
+			return o.del()
 		},
 	}
-)
 
-func del(args []string) error {
-	key := args[0]
+	cmd.PersistentFlags().BoolVarP(&o.IsForceDelete, "force", "f", false, "Force deletion of key and value\nDefault, display a prompt to confirm that you want to delete")
+	cmd.SetOut(globalOption.Out)
+	cmd.SetErr(globalOption.ErrOut)
 
-	param, err := ssmManager.GetParameter(key, false)
+	return cmd
+}
+
+func (o *delOption) del() error {
+	param, err := o.SSMManager.GetParameter(o.Key, false)
 	if err != nil {
-		fmt.Fprintln(ErrWriter, color.YellowString(fmt.Sprintf("WARN: `%s` is not found. Nothing to do.", key)))
+		fmt.Fprintln(o.ErrOut, color.YellowString(fmt.Sprintf("WARN: `%s` is not found. Nothing to do.", o.Key)))
 		return nil
 	}
 
-	if !isForceDelete {
-		fmt.Fprintf(ErrWriter, "Delete `%s` (value: %s) ? (Y/n)\n", key, param.Value)
+	if !o.IsForceDelete {
+		fmt.Fprintf(o.ErrOut, "Delete `%s` (value: %s) ? (Y/n)\n", o.Key, param.Value)
 		scanner := bufio.NewScanner(os.Stdin)
 		for scanner.Scan() {
 			yn := scanner.Text()
@@ -43,18 +73,14 @@ func del(args []string) error {
 			} else if yn == "N" || yn == "n" {
 				return nil
 			} else {
-				fmt.Fprint(ErrWriter, "(Y/n) ?")
+				fmt.Fprint(o.ErrOut, "(Y/n) ?")
 			}
 		}
 	}
 
-	if err := ssmManager.DeleteParameter(key); err != nil {
+	if err := o.SSMManager.DeleteParameter(o.Key); err != nil {
 		return fmt.Errorf("%s\n%s", ErrMsgDeleteParameter, err)
 	}
 
 	return nil
-}
-
-func init() {
-	DelCommand.PersistentFlags().BoolVarP(&isForceDelete, "force", "f", false, "Force deletion of key and value\nDefault, display a prompt to confirm that you want to delete")
 }